delete expériment component with Tailwind and correct the compoentn texte-présentation-page-accueil.jsx without css
