+++ frontend/src/App.css
@@ -1,10 +1,11 @@
+/* Import de Tailwind */
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
 
-/* Import des polices de texte Hind et Expletus.
+/* Import des polices de texte Hind et Expletus. */
 @import url("https://fonts.googleapis.com/css2?family=Hind:wght@300&display=swap");
-@import url("https://fonts.googleapis.com/css2?family=Expletus+Sans:ital,wght@1,400;1,500&family=Hind:wght@300&display=swap"); */
+@import url("https://fonts.googleapis.com/css2?family=Expletus+Sans:ital,wght@1,400;1,500&family=Hind:wght@300&display=swap");
 
 * {
   margin: 0;
@@ -14,14 +15,14 @@
 
 .titre-texte-présentation-page-accueil-tailwind {
   font-family: "Expletus Sans", sans-serif;
-  background-color: #b52030;
+  background-color: #141414;
   color: #f0edee;
   @apply text-center mr-4 ml-4 text-2xl font-bold no-underline;
 }
 
 .paragraphe-texte-présentation-page-accueil-tailwind {
   font-family: "Hind", sans-serif;
-  background-color: #b52030;
+  background-color: #141414;
   color: #f0edee;
   @apply text-center mr-4 ml-4 text-base no-underline;
 }
