+++ frontend/src/App.jsx
@@ -1,5 +1,4 @@
 import TextePrésentationPageAccueil from "./components/TextePrésentationPageAccueil";
-import ExperimentTailwind from "./components/ExperimentTailwind";
 
 import "./App.css";
 
@@ -7,7 +6,6 @@ function App() {
   return (
     <div>
       <TextePrésentationPageAccueil />
-      <ExperimentTailwind />
     </div>
   );
 }
