--- frontend/src/components/TextePrésentationPageAccueil.css
@@ -1,25 +1,0 @@
-/* Style du composant Texte de présentation sur la page d'accueil. */
-
-/* Import des polices de texte Hind et Expletus. */
-@import url("https://fonts.googleapis.com/css2?family=Hind:wght@300&display=swap");
-@import url("https://fonts.googleapis.com/css2?family=Expletus+Sans:ital,wght@1,400;1,500&family=Hind:wght@300&display=swap");
-
-/* Code CSS classique si on n'utilise pas Tailwind.*/
-/* Mettre texte en blanc + police Hind + centré */
-.TextePrésentationPageAccueil {
-  text-align: center;
-  text-decoration: none;
-  color: #f0edee;
-  background-color: #141414;
-  margin: 1em;
-}
-
-.h1-TextePrésentationPageAccueil {
-  font-family: "Expletus Sans", sans-serif;
-  font-size: x-large;
-  font-weight: bold;
-}
-
-.p-textePrésentationPageAccueil {
-  font-family: "Hind", sans-serif;
-}
